refactor graphanalytics
